// heh-ui/src/app/models/service.ts
export interface Service {
  vendor: string;
  title: string;
  category: string;
  tags: Array<string>;
  description: string;
  addressList: Array<string>;
  website: string;
  phones: Array<string>;
  workingHours: string;
  validity: string;
  instagram: string;
  facebook: string;
  vk: string;
}